Rename file: report-head-v2 -> report-head
