+++ src/tests/unit/tests/DetailsView/reports/components/report-head-v2.test.tsx
@@ -2,8 +2,7 @@
 // Licensed under the MIT License.
 import { shallow } from 'enzyme';
 import * as React from 'react';
-
-import { ReportHead } from '../../../../../../DetailsView/reports/components/report-head-v2';
+import { ReportHead } from '../../../../../../DetailsView/reports/components/report-head';
 
 describe('ReportHead', () => {
     it('renders', () => {
